requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.1
boolean.py==4.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cyclonedx-python-lib==7.6.0
defusedxml==0.7.1
distro==1.9.0
feedparser==6.0.11
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-texttospeech==2.17.1
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
license-expression==30.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
numpy==1.26.4
openai==1.42.0
packageurl-python==0.15.6
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
proto-plus==1.24.0
protobuf==4.25.3
py-serializable==1.1.0
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
ratelimit==2.2.1
referencing==0.35.0
requests==2.32.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
schedule==1.2.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.37.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
